newnew 'overview' view and corresponding sidebar
the 'overview' view of section 'db' now includes a tab navigation which enables support for more database objects besides tables

the sidebar for postgresql dialect now looks like that for 'section' tbl while that for mysql dialect looks like that shown in 'home' view i.e. it is a list of database with the currently viewed one selected (displayed as bold).